Move common implementation of fs.use_downloader into fs.tools module
